Use a subfolder "accounts" underneath the config folder
Foreseeing that the config folder will have more things at
the root soon, and willing to separate the important stuff
(private keys) from the rest.

Electrum is similar in this case, but instead of using the
subfolder name "accounts", they use "wallets". I decided to
not use this word because gwallet is supposed to be just one,
which holds many accounts (one per currency, for now).