Fix build in windows (MS.NET)
Previous commit[1] to fix build in Linux switching from .NET
4.5.2 to 4.5 broke the build in Windows with this:

The primary reference "GWallet.Backend.dll" could not be resolved because it has an indirect dependency on the assembly "EdjCase.JsonRpc.Client, Version=1.0.5.0, Culture=neutral, PublicKeyToken=null" which was built against the ".NETFramework,Version=v4.5.1" framework. This is a higher version than the currently targeted framework ".NETFramework,Version=v4.5"

So we switch to 4.5.1 instead of 4.5. This one works for
both Linux and Windows.

[1] https://gitlab.com/knocte/gwallet/commit/a52d98107f0eb2a6d8dbdb42d9a62c4d136af59e