JsonRpcSharp: avoid hangs at first-byte
The previous "sync" algorithm assumed that the first time
it had to start reading, the DataAvailable property of the
stream gave a misleading "False" value. This value was not
misleading, it just took a while until the data was received
by the server and a response started to arrive, but the
approach worked because:
1. The ReadByte() was blocking the thread, but for very
little time.
2. It had never failed (always received data), at least
during all my previous testing.

But today I hit a server that was not sending any response
at all, so the code was getting blocked at the first ReadByte()
call. This teaches me that you cannot really convert an
asynchronous workflow into a synchronous one without having
a timeout mechanism.

So this is what I implemented in this commit. If I hit the
misbehaving server again, at least a timeout now will happen
and an exception will be thrown (and then I will change the
code to handle this gracefully and switch to a different
electrum server).