Backend: fix using accounts generated with previous Nethereum 1.x version
When we switched to using Nethereum 2.0(rc4) [1],
the generated addresses in the json key store files
started having the "0x" prefix (change of behaviour)
so we had to change the backend to expect this.

However, doing this change meant that when dealing
with older json keystores (generated with previous
versions of this wallet) would break. So now with this
change the backend is agnostic and can deal with both
of them.

[1] https://gitlab.com/knocte/gwallet/commit/71d2ca58ffa682c7bfd38abe46c0ef272ffbfab6