Backend,Frontend: don't rely on enum values for serialization
Don't rely on internal representation of types (in this case
an enum, the internal representation being its integer type)
for serialization, otherwise if Backend changes this number
in the future, deserialization would come up with a different
currency!

The indexing strategy for the frontend now is simply take the
order in which the elements come from the reflection functions.

To fix this shit, I had to fix the NewtonsoftJson converter that
I had found on the internet, because it didn't work for my cases.
(Sadly, by doing this I'm sure I've broken other cases, but
since now I'm using Tests for Serialization/Deserialization,
I'll catch those early on; so long as I keep being disciplined
and write tests always for new serialization/deserialization
work.)