Backend: fix caching deserialization
Probably it got broken when making Currency
a DU instead of an enum (in a very recent
previous commit).

I'm not really happy with this solution since
I would prefer the types not be polluted with
serialization-related decorations. Maybe I can
migrate to Newtonsoft.Json.FSharp[1] (not Chiron
because that requires you to pollute your types
too, particularly, with ToJson() and FromJson()
methods).

Fixes https://sentry.io/nblockchain/gwallet/issues/311800628/

[1] https://github.com/haf/Newtonsoft.Json.FSharp