Initial Bitcoin support: allow adding readonly BTC account
The rest of operations are unsupported and will throw exceptions.

The reason to change the way nuget is run is because the
version of nuget that comes from Ubuntu 16.04 doesn't work
when trying to retreive NBitcoin, as it gives this error:
"'NBitcoin' already has a dependency defined for 'Newtonsoft.Json'"
I tried several solutions out there, such as using the option
-self for nuget, however this has two problems:
1) it would need sudo rights to update itself in the system
(this works in gitlab CI but not in general)
2) it can break easily if the nuget team decides to publish a
too bleeding-edge version, e.g.: https://github.com/NuGet/Home/issues/2775

In the end I bundled two target files from https://github.com/kzu/NuGet.Restore
which use curl to download a reasonable version of nuget (4.1.0.2450)
from https://dist.nuget.org/win-x86-commandline/latest/nuget.exe
(which gets downloaded to the .nuget/ subfolder), which works for
me in this system.