XamForms.iOS: fix build switching Linker behaviour
I was getting this compiler error with this Mac:
/Users/andres/Documents/gwallet/src/GWalletFrontendXamForms/iOS/MTOUCH:
Error MT0091: This version of Xamarin.iOS requires the iOS 10.3 SDK
(shipped with Xcode 8.3). Either upgrade Xcode to get the required
header files or set the managed linker behaviour to Link Framework
SDKs Only (to try to avoid the new APIs). (MT0091)
(GWalletFrontendXamForms.iOS)

This Mac has iPhone iOS 10.2 (and XCode 8.2 I think) so I'm
switching this build setting here to make it build with these
old versions.

FIXME: disable this and upgrade, so that I can link all (less
disk space for the app!)