Backend/Bitcoin: rather use NBitcoin's Sign(privKey,Coin[]) method
Rather use NBitcoin's Sign(privKey,Coin[]) method
instead of Sign(privKey,assumeP2sh:bool) which has
now been deprecated[1], as according to Nicolas
Dorier, the segwit way of signing will use the same
abstraction as the Coin[] one, and this deprecated
method requires re-setting the scriptSig property
of the inputs (it can be seen that this commit
removes the need for that).

Sadly this generates bigger transactions and I still
don't know why (see the change in transaction size
estimation in this same commit). The difference is
not so big and we're going to move away from using
Transaction class anyway (in favour of the other
TransactionBuilder approach, which will provide
estimation in itself).

[1] https://github.com/MetacoSA/NBitcoin/commit/d1bd95e29c621ebcd3e2f233274887583184543c