Backend.Portable: introducing new assembly
I need to create this assembly because otherwise I cannot
link the backend from the XamForms library (not compatible;
although this might not be a problem in the near future thanks
to .NET Standard).

The way it works is: this project will simply include all
files from GWallet.Backend as a link, for code reuse.

This is just the first version of the assembly, which only
includes the "Marshalling.fs" file, and which I had to adapt
already to be .NET Portable compatible, via:

https://stackoverflow.com/questions/19718232/assembly-getexecutingassembly-doesnt-exist-in-pcl