Backend,Tests,Frontend: refactor to always substract miner fee in backend
Stop substracting miner fee from amount outside the backend
as it was duplicated logic. This also makes it so that the
unsigned JSON transaction files contain the full amount without
considering the fee, because in the future we may implement some
logic that signs the same transaction many times with different
fees (so the frontend would just determine the maximum fee that
the user would be willing to pay) in order to send transactions
with small fees first, and keep increasing them (i.e. RBF) until
it confirms.