Bitcoin/Backend: support for serializing/deserializing signed BTC tx
Previous commit claimed that it also added support for
serialization of signed BTC transactions but it wasn't
really working (it didn't crash, but the way they were
serialized was causing deserialization not to work because
the type embedded in the metadata contained IBlockchainFee
instead of a concrete Fee type).

This commit fixes this issue, plus finishes the support for
deserialization.

BONUS:
If you look closely at this diff, you might find some
weird match pattern that includes a "box" keyword. If
you want to find out why's this, check this and figure
how much fun I've been having with my friend the F# compiler:
https://fsharpforfunandprofit.com/troubleshooting-fsharp/#FS0008