Backend(.Tests): make FaultTolerance be exception-type aware
From now on it will only retry for the exception type
passed in as a type parameter.

The reason for this is that we may not want to retry for
every exception, but reraise() the ones that are GWallet's
bugs, so we can fix them.

For example:
Fault warning: System.Net.Sockets.SocketException: Connection refused
Fault warning: System.Net.Sockets.SocketException: Network is unreachable
Fault warning: System.NullReferenceException: Object reference not set to an instance of an object
Fault warning: GWallet.Backend.JsonRpcSharp+NoDataReceived: Exception of type 'GWallet.Backend.JsonRpcSharp+NoDataReceived' was thrown.
Fault warning: GWallet.Backend.JsonRpcSharp+NoDataReceived: Exception of type 'GWallet.Backend.JsonRpcSharp+NoDataReceived' was thrown.
Fault warning: System.Net.Sockets.SocketException: Connection timed out

Of the above, there's one that was a GWallet bug,
and given the previous retry-all mechanism, it was
ignored :( so I cannot fix it until I reproduce it
again.