Backend: for archived accounts, check unconfirmed balance, to fix bug
This commit fixes the following bug (in the bitcoin side of things):

Steps to reproduce:
1. Create BTC account A.
2. Create BTC account B.
2. Write down account A's address.
3. Archive account A.
4. Send some BTC to the address you wrote down.
5. Let the payment confirm.
6. Open GWallet: it informs you that an archived account has received funds.
7. You specify to redirect the funds to account B.

Previous results:
After the transaction was broadcasted, and all balances
are refreshed again in GWallet, it reports again that the
archived account A has received funds (because the
transaction to B hasn't confirmed yet), and asks you to
choose an account (again) to redirect funds to.

Current results (after this fix):
After the transaction is sent, hopefully the account will
have an _unconfirmed_ balance of zero when querying the
server again, and it will not prompt you again.