Backend/FaultTolerantClient: support for requiring N consistent items
This is especially useful in an environment in which we have
many servers and don't want just to get the first successful
response, but one which many servers agree on.

The way I have implemented this is refactoring the class
FaultTolerantClient to have two constructors:
* An empty one that works exactly the same as before (just
returning a result as soon as it finds one)
* An extra one that receives an int which determines the
number of consistent results that need to be returned across
all functions to be considered succesful.

The best example for usage in GWallet is Electrum Server querying,
which will make use of this soon.