build(nuget): use -DisableParallelProcessing to try to fix hangs
Since I introduced a new version of nuget (required
by some nuget packages; see 9804b653), the build hangs
usually in gitlab-ci (maybe because of some underlying
Mono problem with HTTP requests, that hopefully is fixed
in a new version... maybe fixed by [1]?).

As a workaround, I'm testing now this flag to see if it
improves things. See the docs here:
https://docs.microsoft.com/en-us/nuget/tools/cli-ref-restore

[1] https://github.com/mono/mono/commit/cd09ddbc55c9866703d43fb8b089978dec2c4ef3