Upgrade Newtonsoft.Json from 9.x to 10.x (& Nethereum from 2rc4 to 2.0.1)
The motivation for upgrading Newtonsoft.Json is because I
was getting a weird failure with v9.x on the JSON-related
unit tests when merging master to this 'bitcoin' branch,
specifically because of bringing the fix for Fsharp.Core[1];
upgrading to 10.x makes it go away (and anyway the 'frontend'
branch already depended on this version because of some
newly-introduced dependency).

The reason for upgrading Nethereum libs is because the old rc4
version still depended on Newtonsoft.Json 9.x, and 2.0.1 is
the first version of Nethereum that depends on 10.x.

[1] https://gitlab.com/knocte/gwallet/commit/ac49dddaea77f313f45704a2b3835b774d7c9725