Backend/JsonRpcSharp(Bitcoin): fix DNS resolution not respecting timeout
It turns out that .NET has a little bug: for HTTP requests,
it correctly adds domain name resolution to the time alloted
for async tasks, but not for TCP requests, which then requires
to do this hack to manually resolve and control the timeout
to resolve. If we don't do this, it could take ages for the
wallet to start in offline mode.