Frontend.XF: general "ignore" cleanup
Ignoring tasks can be dangerous because we could be ignoring
exceptions that happen when they run. Rather let's have this
helper that will double check if an exception happened, and
if it did, it would bring it to the main thread to crash the
app (and consequently to send the error to Sentry, so that we
get notified). This is much better than just leaving the user
with a hanging app.