Backend,Frontend.Console: show unconfirmed balances only if lower
If we have to show only one kind of balance, between
the confirmed and the unconfirmed one we should always
use the former, for safety (so that the user doesn't
assume he has already been paid, and a double-spend
transaction ends up making his incoming transaction
invalid) and usability (so that the user doesn't get
an error, possibly, when trying to spend unconfirmed
transaction outputs).

However, there's one case in which it's safe enough
to use unconfirmed balance, and it also increases
usability as well: when the unconfirmed balance is
lower than the confirmed one. It's not more unsafe
because this would only show different amount when
sending transactions, not for receiving ones. And
it increases usability because after sending an
outgoing transaction and switching to the balance
view again, the user can see the updated balance
immediately (so that he doesn't try to spend in
a subsequent transaction more balance than what he
actually has when the transaction confirms).