Backend,Frontend.Console: support creating accounts from same privKey
This is the first step towards being able to have
one single seed phrase for all currencies.

The Xamarin.Forms frontend will need to consume this
by using this new CreateBaseAccount() method, which
will generate N account files (where N is the number
of currencies that GWallet supports), all derived from
the same private key.

This is better than storing a single encrypted file
for all currencies (even if all of them come from
the same private key) for 2 reasons:
- each currency has its own encryption default format
(in Bitcoin it's WIF, in ETH it's some JSON format
documented somewhere), which each sub-library
(NBitcoin and Nethereum respectively) consumes.
- by having many files, this is essentially a
backup with N-1 copies (in case some of the files
gets corrupted or something); this would require
manual regeneration of the file for now, but it's
better than no backup.

Last but not least, this paves the way for ERC20
token support, because in that case the tokens
will be held in the same ETH|ETC accounts/addresses
(not separate, because to move the tokens the user
will need ETH/ETC gas).