Backend/Bitcoin/Stratum: improve error handling
The exception making GWallet crash should contain the
electrum server name that returned the faulty response.
In order to do this, I had to refactor the error
handling in StratumClient.fs, fixing some code smells
like having to pass the originalRequest json string
as a parameter for the only purpose of including it
in a potential error message. This simplified things
a bit (e.g. the tests).