Backend/Ether: EIP155 signing support (replay attack protection)
So, when I implemented the feature about having a unique
seed private key for all currencies[1], I realised it would
make ETC and ETH currencies have the same public address,
which in turn would be unsafe if the signing method would
not be using any replay attack protection mechanism.

Ethereum and EthereumClassic chains provided ways to
prevent this attack (EIP155[2]), but not all libraries
adopted them. Such is the case of Nethereum, which was
part of this unfortunate group[3]. But I went ahead and
decided to fix this myself, so after couple of days of
work I submitted this PR for review[4].

The change was too disruptive to the eyes of the
maintainer, so based on my implementation, he decided
to take a similar but alternative approach (at least
he kept my tests! hehe). Either way, the final result
is that Nethereum is now compatible with EIP155, since
version 2.3.0 [5], which we adopt in this commit.

[1] https://gitlab.com/knocte/gwallet/commit/147fbe94e361e343814d3af807d2136fb655c935
[2] https://github.com/ethereum/EIPs/blob/master/EIPS/eip-155.md
[3] https://github.com/Nethereum/Nethereum/issues/91
[4] https://github.com/Nethereum/Nethereum/pull/268
[5] https://github.com/Nethereum/Nethereum/commit/647beaff324151c21702cfc68a65f96ff11cbdfc