Preliminary ERC20 token support (i.e. in the form of DAI token, for now)
This work brings native support for ERC20 tokens,
and the first token that is exposed to the UI is DAI.

DAI token is one of my favourites because:
- It's decentralized, in theory (as opposed to one single
centralized actor having full control over it*).
- It has utility (stable value), as opposed to merely
representing "blockchain shares" with the promise of
having a future platform accepting them.

This work could be extended by making this more
parametrizable, e.g. configuring the token address,
and other variables, via config files instead of
being hardcoded in the code.

The upgrade of Nethereum dependency is needed
because we use new ERC20-method estimation APIs
exposed in the last version.

*=Most ERC20 tokens nowadays are just issued by
companies that control (have the private key of
an owner account) the smart contract where the
token lives, most of the time allowing them to
make changes over their issuance and the like.