Backend: fix balance caching since common-seed-private-key feature
Since the common-seed-private-key feature was implemented[1],
balance caching got broken silently because it assumed that
each address can be mapped with one currency. However, since
this feature landed, you get an ETC address which is the same
as your ETH address (which is a side-effect, but what we really
needed to achieve, is that user's ERC20-token address be the
same as their ETH address, to make token transfers more
convenient*).

[1] https://gitlab.com/knocte/gwallet/commit/147fbe94e361e343814d3af807d2136fb655c935

*=E.g. if your DAI address was different than your ETH one,
then everytime you wanted to move your DAI you would need
to transfer some ETH to it in order to have enough gas to
send the transaction (or do tricks with the ERC-20's Approve
function).