Backend,Frontend.Console: fix warning & treat warnings as err in !DEBUG
The compiler caught recently a trailing comma
that I introduced by mistake in [1], which I
fixed in [2]. But running this code in the frontend[3]
made me realize that the fix was not just fixing a
warning but a much bigger problem. In my opinion F#
should treat the lack of an "ignore()" function as
a compiler error, not a warning, in order to catch
these bugs faster.

So I'm fixing a warning I found in Frontend.Console[4]
project, so that I can enable "Warnings as errors",
at least when not running in DEBUG mode (so that it's
easier to work on GWallet, but impossible to release
a new version that has warnings).

[1] https://gitlab.com/knocte/gwallet/commit/08a66bc176a487f3d928b422164bdd70f21f6a62
[2] https://gitlab.com/knocte/gwallet/commit/2b235955bf6c58a581066570afed68f3a9ba8edc
[3] https://gitlab.com/knocte/gwallet/commit/7704418f35f95cbdd573412bed0031cd57514284
[4] /Library/Frameworks/Mono.framework/Versions/5.8.0/lib
/mono/msbuild/15.0/bin/Microsoft.Common.CurrentVersion.targets(5,5):
Warning MSB3836: The explicit binding redirect on
"Newtonsoft.Json, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed"
conflicts with an autogenerated binding redirect.
Consider removing it from the application configuration
file or disabling autogenerated binding redirects. The
build will replace it with:
"<bindingRedirect oldVersion="0.0.0.0-10.0.0.0"
                  newVersion="10.0.0.0"
xmlns="urn:schemas-microsoft-com:asm.v1" />". (MSB3836)
(GWallet.Frontend.Console)