ReadMe.md: add WarpWallet (BrainWallet) to Roadmap
The WarpWallet[1] guys have an open bounty with 20BTC
for a brainwallet generated with a very simple password
(8 alphanumeric characters, e.g. 'b234FEzz') for which
they even provide the salt! And nobody has solved it yet.
So it's safe to assume that if we implement brain-wallet
style private key generation with at least 10 characters
minimum and some other requirements (must include numbers
and letters, must include uppercase and lowercase, must
include punctuation) and a stronger salt (not only email
but also birthdate and ID no.?), it would be safe to
assume that we have a strong brainwallet.

Obligatory quoting this too: [2].

[1] https://keybase.io/warp
[2] https://xkcd.com/936/