Upgrade to FSharp.Core 4.3.4 (Version=4.4.3.0)
The MissingMethodException from FSharp.Data explained in
a recent commit message[1] reappeared after I fixed the
build[2] (broken by that same first commit).

The correct fix is to actually upgrade FSharp.Core to
the last version. However, doing this alone would make
a FileNotFoundException appear[3] because the upgrade
to this version makes this change to the .fsproj:

     <Reference Include="FSharp.Core">
-      <HintPath>..\..\packages\FSharp.Core.4.1.0.2\lib\net45\FSharp.Core.dll</HintPath>
+      <HintPath>..\..\packages\FSharp.Core.4.3.4\lib\net45\FSharp.Core.dll</HintPath>
+      <Private>False</Private>
     </Reference>

We then fix this problem by enabling LocalCopy on the
dependency[4], represented more clearly by this diff
below (because somehow upgrading to 4.3.4 would make
this <Private> shit appear):

     <Reference Include="FSharp.Core">
       <HintPath>..\..\packages\FSharp.Core.4.3.4\lib\net45\FSharp.Core.dll</HintPath>
-      <Private>False</Private>
     </Reference>

[1] https://gitlab.com/knocte/gwallet/commit/1f19ae015a7d24f3a473008804a295eb4c05076a
[2] https://gitlab.com/knocte/gwallet/commit/273d9544139da79f73ab476a84daf5c316e69773
[3] https://stackoverflow.com/questions/49584528/4-6-1-f-app-throwing-filenotfoundexception-fsharp-core
[4] https://stackoverflow.com/a/16974650/544947