Frontend.XF(,.Gtk): fix clipboard copy
QR code was not showing because:
1) the Gtk frontend project was not referencing the zxing
libs, so it was crashing -> fixed in this commit
2) some unknown reason about visualization in the GTK platform

As a workaround to (2), I'm also limiting this frontend
to just copying the address to the clipboard instead of
showing the ReceivePage (because no QR code would be shown
for now in the latter). However, this clipboard feature
was not working either because Xamarin.Plugins.Clipboard
didn't support this platform, something I fixed by
contributing to its repository in github[1], and was
published in version 2.2.1 of it, which I'm bringing here.

[1] https://github.com/stavroskasidis/XamarinClipboardPlugin