Backend: increase configuration's default network timeout
This timeout was fundamental before FaultTolerantClient
was parallel, because thanks to it being low, it wouldn't
hang very long for faulty servers until the next one was
retried. However, there's no need for this timeout to be
short anymore since FaultTolerantClient has parallelism
capabilities, because the most responsive server would be
the one that makes its wait to be unblocked.

This also helps in scenarios where the network is slow,
such as countries with poor internet connections, which
would always timeout before this change.