Backend/UtxoCoin: upgrade our Electrum software&protocol versions
There are some servers that started rejecting 0.10 clients,
which we fixed in the previous commit to be ignored[1].
However, by looking at the last protocol updates[2], we
can just claim to be 1.0 compatible (instead of 0.10)
because the methods deprecated in 1.0 were not being used
by us.

This way, we have more servers available to server our requests,
instead of ignoring some.

This commit also finishes a TODO item to test for this,
so that if it ever happens again, we would catch it in
our integration tests (which is better than just ignoring
from the client perspective and never know it).

Still, the tests require to be run (either manually by
a developer, or by CI as a result of an unrelated commit),
so I'm still adding a FIXME to report this on sentry.

[1] https://gitlab.com/knocte/gwallet/commit/8d716b786bf891b03069a1f216c37061bc9f58c5
[2] https://electrumx.readthedocs.io/en/latest/protocol-changes.html