Backend: make FaultTollerantParallelClient's parallelization kick ass
This is the last ice on the cake because it allows
the client to start new tasks whenever there were
failures (or successes, to keep increasing the result
count towards consistency), in order to always have
N requests in flight (N being the parameter named
numberOfMaximumParallelJobs in its ctor).

I'm of course adding 2 unit tests that demonstrate this
neat behaviour. I managed to implement this especially
thanks to the awesome Seq.splitInto function, which I
explain with a test below:

    [<Test>]
    let ``testing``() =
        let someList = [0;1;2;3;4;5;6]
        let someSeq = Seq.ofList someList
        let someBuckets = Seq.splitInto 3 someSeq |> List.ofSeq
        Assert.That(someBuckets.Length, Is.EqualTo(3))
        Assert.That(someBuckets.[0].Length, Is.EqualTo(3))
        Assert.That(someBuckets.[0].[0], Is.EqualTo(0))
        Assert.That(someBuckets.[0].[1], Is.EqualTo(1))
        Assert.That(someBuckets.[0].[2], Is.EqualTo(2))
        Assert.That(someBuckets.[1].Length, Is.EqualTo(2))
        Assert.That(someBuckets.[1].[0], Is.EqualTo(3))
        Assert.That(someBuckets.[1].[1], Is.EqualTo(4))
        Assert.That(someBuckets.[2].Length, Is.EqualTo(2))
        Assert.That(someBuckets.[2].[0], Is.EqualTo(5))
        Assert.That(someBuckets.[2].[1], Is.EqualTo(6))
        ()