Backend/Marshalling: fix little FIXME
(This was originally committed[1] in the frontend branch
because the way to get the version had to be changed
there for reasons related to .NETStandard/Portable
support. But it should have been committed in the
master branch and then backported to the frontend
branch, otherwise it makes the codebase more prone
to merge conflicts, like it's happenning now. So I'm
backporting this to maybe improve the situation for
further merges.)

[1] 0c43a6fa7e143ff9abde7c911261bdb5e2e91314