Frontend.XF: overhaul main balances page
Instead of having SEND and RECEIVE buttons for each
currency row cluttering the UI, we move this functionality
to the ReceivePage, where it shows the QR code and where
we place a SEND button.

Also this major UX/UI update covers including FIAT amounts
(which was not possible before because of FSharp.Data being
unfriendly to .NETStandard[1]).

[1] 06e54c07e9f15d80e7545aead9993abddf08c2c0