Backend/UtxoCoin(,Tests): rework some electrumServer errors
IncompatibleServer shouldn't mean "connection unsuccesful"
strictly speaking, so we break the inheritance there, and
the effect of this is:
- We need to expand the try-with block in UtxoCoinAccount
to also make these IncompatibleServerExceptions translate
into ElectrumServerDiscarded errors.
- We don't need anymore the ServerTooNew assert in the
tests, because if it happens, it will simply not get
swallowed by the try block anymore, and bubble up the
exception and make the tests fail.