Backend/UtxoCoin: for insufficient funds edge case, earlier assertion
In the case the amount plus the estimated fee surpasses
the balance in BTC or LTC, an assertion was happening at
transaction signing time, from NBitcoin's Transaction.Check()
verification method, whose output was "NegativeOutput".

Let's make this assertion be in our backend code (instead
of relying on NBitcoin) and happening much earlier, so that
we can analyze it later, in order to find a better bugfix.