Backend/UtxoCoin: rework miner fee substraction to happen earlier
This all goes inline with the previous commit in which
we made an assertion happen earlier: now when it happens
that fee+amount>balance, we don't even prompt the user
for accepting the fee, and fail even faster. Because now
we substract the miner fee from the transaction draft at
fee-estimation-time and not at signing time.

To be able to do this refactoring we needed to shuffle
some code around, such as moving the private function
SubstractMinerFeeToTransactionDraft a bit further up in
the file where it lives.