Backend,Frontend.Console: raise InsufficientBalanceForFee
There were 3 cases depending on the amount to be sent by the user:
1. All balance: in this case the fee was deducted from the amount.
2. Some balance sent, some balance left: the fee was deducted from
the balance remaining after substracting the amount to be sent.
3. Some balance sent, not enough balance left for the fee: this
was throwing an error, depending on the currency:
 a) BTC,LTC: an assertion when substracting the fee from the tx
 b) ETH,ETC: an InsufficientFunds exception caught from an RPC
 response from the server.
Now with this commit, in both cases (a) and (b) we raise a new
kind of exception called "InsufficientBalanceForFee".

An exercise for the reader(?): how to do this with DAI? Later,
later... or my head will explode.