Backend(UtxoCoin): upgrade protocol version from 1.0 to 1.1
Some integration tests started failing [1], presumably
because some Litecoin servers started to be picky about
the version of the clients they serve too. I've looked
at the specs of the changes between 1.0 and 1.1 [2], and
it's mainly addition of methods, removal of methods that
we were not using, or changes in 'server.version' which I
simply accommodate with, in this commit.

[1] https://gitlab.com/knocte/gwallet/-/jobs/71249985
[2] https://electrumx.readthedocs.io/en/latest/protocol-changes.html