Backend: enable consistency checking in ETH response management
After we introduced 2 more public ETH servers[1],
we can now enable the consistency checking: if 2
servers respond with X and 1 server responds with Y
then we take X.

Problem about this was the fact that Nethereum had
a little bug[2] in the HexBigInteger class, which
Juan fixed right away[3], so I'm bringing the new
Nethereum version from CI to make this work (pulling
binaries from MyGet). For this I had to create a
custom NuGet.config to point to these custom packages.
Whenever there's a new stable Nethereum release, it
will be better to switch to that (especially for a
stable gwallet version).

[1] https://gitlab.com/knocte/gwallet/commit/66862cc8bde29c1340b631d70ae054cf1f9e3288
[2] https://github.com/Nethereum/Nethereum/issues/359
[3] https://github.com/Nethereum/Nethereum/commit/2d1cb59461d208b8382efd3875f41fb783ae26c2