Backend(,Tests),Frontend.Console: add immediate balance subtraction
This is an approach that I'm not 100% happy about (not even
50% maybe... ok, this is a huge HACK). Thing is, to implement
this properly I would need to query each currency's servers
independently to check the number of confirmations, which is
tricky and takes time to develop (and maybe it's a flaky way
after all...).

The purpose of this change is to introduce some caching of
the values of transactions that have just been sent by
gwallet so that it can show the substracted balance immediately
(even before the transaction has had any confirmation at all).

The tricky part of this approach is not really so much the
algorithm underneath (actually this is tricky as well, check
some changes in the Caching module...), but the hardest part
is the invalidation of the cache, which involves some lists
combinations algorithm in case the user sends more than one
transaction and the balance comes updated at once taking in
account many of them (yes, this would be fragile if all those
transactions have the same amounts...).

Anyway, at some point I see this code going away in favour
of what I said earlier: tracking number of confirmations of
each transaction.

TODO for this current hacky approach: invalidate outgoing
transactions in the cache based on the date (e.g. if older
than 24h discard?) just in case it never confirms.