Backend(,Tests),Frontend.Console: move amount formatting to Backend
This text formatting/presentation logic is going to be reused
by the other frontends so it's better to move it to the backend.

Besides, it's getting changed so often recently that it's better
to add tests to make sure we don't regress and make it more clear
what do those (upcoming) changes achieve from now on.