Backend(Formatting): bring back trailing zeroes, but only for fiat
The change to remove trailing zeros [1] was actually
only meant for crypto amounts (so that it shows "0" instead
of 0.00000, which looks better). But it turns out that for
fiat amounts such as USD, it's always more clear to show
the decimals (because they are only 2) for clarity.

Note, in this change we're also changing some previous
tests so that they only test thousand separation and not
trailing zeroes.

[1] 220c06279928cbddf5dd616f644bd425c7afd7c9