Frontend.XF: be more navigation friendly
If we use a nav page as the receive page, we don't need
the ad-hoc "Go back" button, because we have a built-in
one at the top (except in Mac, where there's a bug and
we add a workaround).

This change will also become useful because to achieve
this we need to stop using modals, which will allow us
to remove pages in the stack (you cannot remove modal
pages in the stack in current XamForms3.x, maybe it's
a bug?). Look out for upcoming next commit (or next
after next) that will make use of .RemovePage()...