Backend,Frontend.XF: 2nd part of cold storage support
This allows showing a QR code when sending a transaction
from a cold-storage (readonly) account.

Last part for cold storage will be to be able to scan
this QR code.

Two clarifications about this commit:
a) The serialized transaction json was too big to fit in
a QR code, so we had to compress it.
b) We renamed passphrase to password because since we use
WarpWallet as a brain-wallet private-key generation, the
decryption secret is a "password" and the seed secret is
a "passphrase".