Frontend.XF: start a bit earlier to refresh cold balances
So the LoadingPage->BalancePage was designed in a way that
would prevent checking the balances of cold-storage addresses
get in the way of checking the balances of hot-storage
(just using cache for the former and checking real servers
for the latter). The problem about this is that then, after
the balances were shown for the first time (after the
'Loading...' page disappears) it would take a whole cycle
(currently configured as 1 minute) to start refreshing the
cold storage balances.

Now this changes it to still not check cold-storage balances
in the LoadingPage, but start checking them right away after
we show the updated hotstorage ones.