Frontend.XF(.*): set of paranoid measures to try to fix big QRcode scan
I'm having a lot of problems when testing the recently committed
feature (previous commit) of scanning a QR code of a transaction
proposal (cold storage mode). It seems larger QR codes cannot be
scanned or are sometimes scanned incorrectly.

To try to countermeasure this problem, this commit:
- Upgrades ZXing.Net.Mobile* deps to the last version.
- Adds options to only expect QR codes (not other kind
of bar codes) when scanning.
- Adds more paranoid options such as TryHarder (no idea
what this does but sounds harmless and helpful) and
UseNativeScanning (which unfortunately a.t.m. seems to
do nothing...[1] but whatever).

But most importantly:
- Adds an alert message for the user to take in account that
he has to fit the whole QR code as much as he can to fit the
screen almost entirely (maybe we need a rewording of this to
make this clear?). Doing this I've noticed the false positives
are less.
- Prevents the false positives from happening at all by adding
a HACK preventing low-number-of-character results to go through
(which is what happens when this bug kicks in, as tested by me
and reported in the associated bugs[2][3]).

[1] https://github.com/Redth/ZXing.Net.Mobile/issues/386
[2] https://github.com/Redth/ZXing.Net.Mobile/issues/624
[3] https://github.com/micjahn/ZXing.Net/issues/135