Backend,Frontend.Console: delay side-effects of creating accounts
Instead of including the creating of the config files of the accounts
in the asynchronous process of the creation of each account, we just
leave them in a state called "ConceptAccount" created in memory, that
we write later (to the config folder) once all conceptAccounts have
been created.

This is to avoid a possible bug that could happen if the creation
of the accounts in a mobile platform (e.g. Android) is too slow, and
the user switches (on purpose, or by unintentionally through screensaver)
the app to be in the background. The result of this, before this commit,
could very likely be that some accounts were created and some were not
so the user would only see, for example, 2 currencies instead of 5.

This refactoring of course is not a full-proof fix (because this switch
could still happen when the config files are being written), but it
makes it much less likely, because now all config files are written
at roughly the same time (still sequentially, but much closer in
time together than before).