Backend,Frontend.*: if no imminent payment, higher interval for timer
In each balance refresh, we can check if there's an imminent
payment coming, by comparing the confirmed and the unconfirmed
balance of each account. If the unconfirmed amount is higher,
we have an incoming payment that may confirm soon, so it's more
important to refresh balances in a more frequent way when this
happens.

If there's no imminent payment about to be confirmed, then there's
no need to refresh balances so often (in 5 minutes time no payment
is going to be received && confirmed anyway*), so after this commit
we stop checking balances every minute, which drains battery and,
most importantly, creates too many threads that may clog the app
(although we need a better fix for this latter issue, probably using
cancellation tokens).

* at least with the number of confirmations that we require nowadays