build: beginnings of a --paranoid mode
NuGet packages are binary. This is not good from a trust
point of view. In the same way you shouldn't use a crypto
wallet that is not opensource, you should not use a crypto
wallet whose dependencies are binary-only. This is why, at
least for our Linux builds, we should introduce a --paranoid
building mode that retreives everything from source instead
using Nuget at compile time.

The beginning of this mode is just introducing the submodule
for our ZXing.Net.Mobile.Forms dependency, because we now
have the need to debug it.