Backend.Tests: make NUnit tests able to run in macOS
This seems like the upside-down world, but Mono's installer
for macOS brings an older version of nunit-console (2.4.8)
than what comes in Ubuntu (2.6.4). And the consequence of
this is that this old version of NUnit cannot run tests
written in an idiomatic F# way (with modules and lets) so
no tests were run at all when the suite was launched in
macOS.

A workaround to this problem is to simply rewrite the tests
in a more OOP way like I do in this commit. Yes, it's ugly
but it's cross-platform without the need to download any
runner binary from the internet.